requirements.txt
playwright
pandas
openai
python-dotenv
duckdb
pyarrow
tqdm
twilio
discord.py